Refactor GUI classes into separate modules
